requirements.txt
streamlit
easyocr
Pillow
numpy
pandas
torch
opencv-python-headless
